Modified channel model in HD test for compatibility with older data
At some point default random failure for RAP channel was raised to 0.1 from 0.05. This is more in keeping with observed packet failure rates.
However, the Oceans 2023 paper used older code for the HDC mode table, which had a random failure rate of 0.05, so the script for creating the data was modified to deliberately set random failure of RAP to 0.05